Manage Product Inventory

+++ cart/src/main/java/com/ecom/domain/CartItem.java
@@ -8,8 +8,6 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Positive;
 import java.time.LocalDate;
 
 @Data
@@ -25,7 +23,7 @@ public class CartItem {
     @NotBlank(message = "cart.productId must be present")
     private String productId;
     @Min(value = 1, message = "cart.productQuantity must be more than 0")
-    private int quantity;
+    private long quantity;
     private LocalDate createdAt = LocalDate.now();
     private LocalDate modifiedAt = LocalDate.now();
 }
